renamed user-order to order and started cleaning up order code
