+++ client/routes.js
@@ -1,25 +1,22 @@
+'use strict'
+
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
-import { BrowserRouter as Router, Route, withRouter } from 'react-router-dom'
+import { Route, withRouter } from 'react-router-dom'
 import PropTypes from 'prop-types'
-import ProductForm from './components/product-form'
 import NewProductForm from './components/new-product-form'
 import EditProductForm from './components/edit-product-form'
 import Navbar from './components/navbar'
 import { Login, Signup } from './components/auth-form'
-import Order from './components/user-order'
+import Order from './components/order'
 import { me, logout } from './store/user'
 import CategorySelector from './components/category-selector'
-import { Home, ProductList } from './components'
+import { Home } from './components'
 import SelectedProduct from './components/selected-product'
 import { getOrderServer } from './store/order';
 import EditReviewForm from './components/edit-review-form'
-//import { me } from './store'
 
 
-/**
- * COMPONENT
- */
 class Routes extends Component {
   constructor(props) {
     super(props)
@@ -28,10 +25,6 @@ class Routes extends Component {
         this.props.getOrder(this.props.user.id)
       })
   }
-
-  // async componentDidMount() {
-  //   await this.props.loadInitialData()
-  // }
 
   render() {
     return (
@@ -52,14 +45,8 @@ class Routes extends Component {
   }
 }
 
-
-/**
- * CONTAINER
- */
 const mapState = (state) => {
   return {
-    // Being 'logged in' for our purposes will be defined has having a state.user that has a truthy id.
-    // Otherwise, state.user will be an empty object, and state.user.id will be falsey
     user: state.user,
     isLoggedIn: !!state.user.id
   }
@@ -72,13 +59,8 @@ const mapDispatch = (dispatch) => {
   }
 }
 
-// The `withRouter` wrapper makes sure that updates are not blocked
-// when the url changes
 export default withRouter(connect(mapState, mapDispatch)(Routes))
 
-/**
- * PROP TYPES
- */
 Routes.propTypes = {
   loadInitialData: PropTypes.func.isRequired,
   isLoggedIn: PropTypes.bool.isRequired
